+++ packages/core/onboarding/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/onboarding
 
+## 6.1.15
+- [patch] [6855bec](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6855bec):
+
+  - Updated internal use of ModalDialog to use new composition API
+
 ## 6.1.14
 - Updated dependencies [58b84fa](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/58b84fa):
   - @atlaskit/analytics-next@3.1.2